Fix reprocess job: fractional skill years crash + incremental persistence + batch-skip flag
- Skill.Years was *int but LLM sometimes returns fractional years (e.g.
  0.3 for a skill used ~4 months), causing json.Unmarshal to fail and
  discard the ENTIRE extraction for that CV. Changed to *float64
  (downstream already reads years_of_experience as float64 and
  truncates for display).
- Reprocess job previously extracted ALL items first, then applied
  (wrote to DB) all of them at the end -- meaning a mid-run restart lost
  all progress with zero candidates created. Now applies each successful
  extraction (graph/candidate/embedding) immediately, item by item, so
  progress survives interruptions.
- Add GROQ_BATCH_DISABLED=true env var to skip the pointless
  submit-then-403 round trip when the account plan doesn't support
  Batch API (applies to reprocess job, reprocess_cvs CLI, and bulk
  upload routing).

+++ internal/llm/service.go
@@ -62,11 +62,11 @@ type Candidate struct {
 }
 
 type Skill struct {
-	Name           string  `json:"skill"`
-	Proficiency    string  `json:"proficiency"`
-	Years          *int    `json:"years"`
-	Confidence     float64 `json:"confidence"`
-	NormalizedFrom string  `json:"normalized_from,omitempty"`
+	Name           string   `json:"skill"`
+	Proficiency    string   `json:"proficiency"`
+	Years          *float64 `json:"years"` // LLM sometimes returns fractional years (e.g. 0.3) — must not be *int or JSON unmarshal fails and discards the whole extraction
+	Confidence     float64  `json:"confidence"`
+	NormalizedFrom string   `json:"normalized_from,omitempty"`
 }
 
 type Company struct {
